completing migration table

+++ database/migrations/2024_05_01_115511_create_badges_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('badges', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('student_id');
+            $table->unsignedBigInteger('student_id');
             $table->string('title');
             $table->text('description');
             $table->timestamps();
